refactor: centralize outcome processing
